[CEN-1448] refactor and increase logging
