+++ src/main/java/it/gov/pagopa/rtd/ms/rtdmsingestor/event/EventHandler.java
@@ -1,11 +1,13 @@
 package it.gov.pagopa.rtd.ms.rtdmsingestor.event;
 
+import it.gov.pagopa.rtd.ms.rtdmsingestor.model.BlobApplicationAware;
 import it.gov.pagopa.rtd.ms.rtdmsingestor.model.EventGridEvent;
-import it.gov.pagopa.rtd.ms.rtdmsingestor.service.BlobApplicationAware;
+import it.gov.pagopa.rtd.ms.rtdmsingestor.service.BlobRestConnector;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import lombok.Getter;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
@@ -16,6 +18,9 @@ import org.springframework.messaging.Message;
 @Configuration
 @Getter
 public class EventHandler {
+
+  @Autowired
+  BlobRestConnector blobRestConnector;
 
   /**
    * Constructor.
@@ -28,8 +33,8 @@ public class EventHandler {
     return message -> message.getPayload().stream()
         .filter(e -> "Microsoft.Storage.BlobCreated".equals(e.getEventType()))
         .map(EventGridEvent::getSubject).map(blobApplicationAware::init)
-        .map(blobApplicationAware::download).map(blobApplicationAware::open)
-        .collect(Collectors.toList());
+        .map(blobRestConnector::download).map(blobRestConnector::open)
+        .map(blobRestConnector::produce).collect(Collectors.toList());
   }
 
 }
